% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCustomSpatialUnits.R
\name{addCustomSpatialUnits}
\alias{addCustomSpatialUnits}
\title{Add custom spatial units to input data for thematic indices}
\usage{
addCustomSpatialUnits(theme, NIObject = NULL, importData = NULL)
}
\arguments{
\item{theme}{character. Argument specifying which thematic index
should be produced. Required when OutputType = "ThematicIndex". For currently
supported thematic indices, see documentation of listIndicators_thematicIndex().}

\item{NIObject}{list containing assembled NiObject.}

\item{importData}{list containing imported data.}
}
\value{
imported dataset or assembled NiObject with custom spatial units.
}
\description{
A subset of thematic indices use their own specific spatial units.
At present, this is the case for thematic indices "Acidification", "Eutrophication",
"AlpinePasserines", "CoastalSeabirds", and "PelagicSeabirds".
}
\details{
For the first three, custom spatial units get added to the data right after
import and the argument importData needs to be provided.
For the seabird indices, custom spatial units get added to the assembled
NiObject and the argument NIObject needs to be provided.
}
