% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCustomNI.R
\name{calculateCustomNI}
\alias{calculateCustomNI}
\title{Calculate custom index based on data in the Nature Index Database}
\usage{
calculateCustomNI(
  ecosystem = NULL,
  indicators = NULL,
  theme = "None",
  dropInd = NULL,
  username,
  password,
  KeyIndicators,
  KeyWeight,
  AreaWeights,
  TrophicWeights,
  NAImputation,
  years,
  OutputType,
  funArguments = NULL,
  Diagnostics,
  TestRun,
  norwegianNames = TRUE,
  saveSteps = TRUE
)
}
\arguments{
\item{ecosystem}{character. The ecosystem for which to calculate the index.
Optional argument required when OutputType = "NatureIndex" or
"EcologicalCondition". Can be one of c("Skog", "Fjell", "Våtmark",
"Åpent lavland", "Ferskvann", "Kystvann", "Hav"). Note that only the first
three are relevan for OutputType = "EcologicalCondition" so far.}

\item{indicators}{character vector containing Norwegian names of indicators
to include in the custom index. Optional argument that has to be provided
when calculating thematic or fully custom indices.}

\item{theme}{character. Optional argument specifying which thematic index
should be produced. Required when OutputType = "ThematicIndex". For currently
supported thematic indices, see documentation of listIndicators_thematicIndex().}

\item{dropInd}{a vector containing integer IDs for indicators to drop.}

\item{username}{character. Username for the Nature Index database.}

\item{password}{character. Password for the Nature Index database.}

\item{KeyIndicators}{logical. If TRUE, applies key indicator weighting, i.e.
the indicators categorized as key indicators will together make up a
proportion of the index equal to "KeyWeight".}

\item{KeyWeight}{numeric, between 0 and 1. Proportion of the index that is
allocated to key indicators. Only in effect if KeyIndicators = TRUE.}

\item{AreaWeights}{logical. If TRUE, weights indicators according to area.}

\item{TrophicWeights}{logical. If TRUE, weighs indicators according to
trophic group.}

\item{NAImputation}{logical. If TRUE, imputes missing values in dicator data
using MICE (Multivariate Imputation by Chained Equations). If FALSE, ignores
missing values.}

\item{years}{integer vector specifying years for which to calculate index.
Note that at present, data in the Nature Index database is only available for
years 1990, 2020, 2010, 2011, 2012, 2013, 2014, and 2019.}

\item{OutputType}{character. The type of output the workflow will create,
here an optional argument. Can be one of
c("NatureIndex", "EcologicalCondition", "ThematicIndex", "CustomIndex").}

\item{funArguments}{a list of arguments to pass on to downstream functions
from NIcalc. Required for output types "ThematicIndex" and "CustomIndex".
Output of listFunctionArguments().}

\item{Diagnostics}{logical. If TRUE, calculates diagnostics for the dataset
prior to imputation. If FALSE, skips diagnostics calculation.}

\item{TestRun}{logical. If TRUE, runs a shorter version of the workflow with
only 10 iterations for imputation and calculation. If FALSE, performs a full
run with 1000 iterations instead.}

\item{norwegianNames}{logical. If TRUE (default), data and results are returned
with Norwegian indicator and ecosystem names.}

\item{saveSteps}{logical. If TRUE (default), saves results at each step of
the workflow as .rds files into the working directory.}
}
\value{
a list of lists containing general information on the calculated
index ("indexInfo"), the raw dataset downloaded form the Nature Index database
("importData"), the formatted input dataset underlying index calculations
("InputData"), results of dataset diagnostics ("Diagnostics", only if
Diagnostics = TRUE), the imputed values for missing data points ("NAImputes,
only if NAImputation = TRUE), and the calculated custom Index ("CustomIndex").
}
\description{
This function is a wrapper for downloading select data from the Nature Index
database and calculating a choice index from it according to the Nature Index
methodology.
}
\details{
The steps that the function goes through are as follows:
\enumerate{
\item Data download via NIcalc::importDatasetApi
\item Assembly of input data for calculations via NIcalc::assembleNiObject
\item Optional: run dataset diagnostics via NIcalc::imputeDiagnostics
\item Optional: impute missing data via NIcalc::imputeData (which uses mice::mice)
\item Index calculation via NIcalc::calculateIndex
}

There are four groups of indices that can be calculated using this function:
Nature Index for any ecosystem (forest, mountain, wetlands, open lowland,
freshwater, coast, and ocean), modified Nature Index for ecological condition
assessments (for ecosystems forest, mountain, and wetlands), a selection of
pre-defined thematic indices (see listIndicators_thematicIndex for supported
thematic indices), and fully customized indices.

The function allows a high degree of control over the calculation via a range
of control arguments. That way, a user can choose the indicators and years
to include, whether and how to perform indicator weighting, whether or not to
impute missing values, etc.

Running this function requires valid user credentials for the Nature Index
database (https://naturindeks.nina.no/). Credentials may be requested from
NINA by contacting the project manager for the Nature Index. Currently
Chloé R. Nater: chloe.nater(at)nina.no.
}
