% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomapNI.R
\name{geomapNI}
\alias{geomapNI}
\title{Map index summary statistics to geospatial shapes}
\usage{
geomapNI(
  Index,
  year,
  OutputType,
  ecosystem = NULL,
  theme = NULL,
  awBSunit = NULL,
  shapeLibraryPath
)
}
\arguments{
\item{Index}{a list ontaining all information on the custom index. Object
"CustomIndex" in the output of calculateCustomNI().}

\item{year}{integer. Which year to map data for.}

\item{OutputType}{character. The type of index to be plotted. Can be one of
c("NatureIndex", "EcologicalCondition", "ThematicIndex", "CustomIndex").}

\item{ecosystem}{character. The ecosystem of the index. Optional argument
required when OutputType = "NatureIndex" or "EcologicalCondition". Can be one
of c("Skog", "Fjell", "Våtmark", "Åpent lavland", "Ferskvann", "Kystvann",
"Hav"). Note that only the first three are relevant for OutputType =
"EcologicalCondition" so far.}

\item{theme}{character. Optional argument specifying which thematic index
should be plotted. Required when OutputType = "ThematicIndex". For currently
supported thematic indices, see documentation of listIndicators_thematicIndex().}

\item{awBSunit}{character. Optional argument that needs to be provided when
OutputType = "CustomIndex" to determine which spatial units should be used
for plotting. Can be any of c("Fjell", "Skog", "Våtmark", "Åpent lavland",
"Ferskvann"), but note that the outcome is the same for all terrestrial
ecosystems.}

\item{shapeLibraryPath}{character. Path to the folder in which the shapefiles
are deposited.}
}
\value{
an sf object containing geomapped summary statistics for the index in
the specified year.
}
\description{
This function links a range of summary metrics for the index to the relevant
geospatial information. Metrics include median, spread of the confidence
interval (= a measure of uncertainty), mean, standard deviation, and
statistical displacement (an artifact resulting from indicator scaling that
may affect the index estimates.)
}
\details{
Geomapping requires shapefiles for the spatial units used by the Nature
Index for the different ecosystems and the different thematic indices.
These are deposited in NINA's internal file system under
"P:/41201611_naturindeks_2021_2023_vitenskapelig_ansvar/Shapefiles".
Alternatively, the files can be downloaded in JSON format from the NI
database API here: https://ninweb08.nina.no/NaturindeksAPI/index.html. Files
downloaded from here first need to be converted into shapefiles.
}
