% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMergeDataset_CoastOcean.R
\name{importMergeDataset_CoastOcean}
\alias{importMergeDataset_CoastOcean}
\title{Import merged datasets for coast and ocean ecosystems}
\usage{
importMergeDataset_CoastOcean(
  ecosystem_part,
  ecosystem,
  username,
  password,
  year,
  norwegian,
  refYearCode
)
}
\arguments{
\item{ecosystem_part}{character vector of length two with the names of the
sub-ecosystems to combine. c("Kystvann-bunn", "Kystvann-pelagisk") for coast and
c("Havbunn", "Hav-pelagisk") for ocean.}

\item{ecosystem}{character. The ecosystem for which to import a merged data
file. Can be either "Kystvann" (coast) or "Hav" (ocean).}

\item{username}{character. Username for the Nature Index database.}

\item{password}{character. Password for the Nature Index database.}

\item{year}{integer vector specifying years for which to import data.
Note that at present, data in the Nature Index database is only available for
years 1990, 2020, 2010, 2011, 2012, 2013, 2014, and 2019.}

\item{norwegian}{logical. If TRUE (default), data and results are returned
with Norwegian indicator and ecosystem names.}

\item{refYearCode}{integer. The year to use as reference year, where 0 =
reference year as set in the database.}
}
\value{
a list containing a combined dataset for ecosystems coast or ocean.
}
\description{
This is a helper function for importing data for the marine ecosystems "Kyst"
(coast) and "Hav" (ocean).
Data for these ecosystems needs to be imported stepwise as they are stored in
the Nature Index database as two datasets each: one for seafloor and one for
open water (pelagic). This function imports both data sets and combines them
into one for further analysis.
}
