% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFunctionArguments.R
\name{listFunctionArguments}
\alias{listFunctionArguments}
\title{List additional function arguments for thematic and custom indices}
\usage{
listFunctionArguments(
  OutputType,
  theme = NULL,
  predefNIunits = NULL,
  indexType = NULL,
  part = NULL,
  total = NULL,
  partOfTotal = NULL,
  awBSunit = NULL
)
}
\arguments{
\item{OutputType}{character. The type of output the workflow will create. Can
be "ThematicIndex" or "CustomIndex". This function is currently not needed
for OutputType = "NatureIndex" and OutputType = "EcologicalCondition".}

\item{theme}{character. Optional argument specifying which thematic index
should be produced. Required when OutputType = "ThematicIndex". For currently
supported thematic indices, see documentation of listIndicators_thematicIndex().}

\item{predefNIunits}{vector of length 3 containing logical objects named
"allAreas", "parts", and "counties". Optional argument, only required when
OutputType = "CustomIndex".}

\item{indexType}{character. Optional argument passed on to
NIcalc::assembleNiObject specifying whether input data should be structured
according to ecosystem (options "ecosystem", "thematicPerEcosystem") or not
(option "thematic"). Needs to be specified only for fully custom indices.}

\item{part}{character one of c("ecosystem","marine","terrestrial"). Optional
argument passed on to NIcalc::assembleNiObject specifying the numerator
variable in criterion for selecting BSunits.}

\item{total}{character one of c("ecosystem","marine","terrestrial"). Optional
argument passed on to NIcalc::assembleNiObject specifying the denominator
variable in criterion for selecting BSunits.}

\item{partOfTotal}{numeric, between 0 and 1. Optional argument passed on to
NIcalc::assembleNiObject which can be used to restrict the input data set to
BSunits satisfying: part/total > partOfTotal.}

\item{awBSunit}{character. Optional argument passed on to
NIcalc::calculateNIunitWeigth via NIcalc::calculateIndex. Specifies which
ecosystem's BSunits should be used to calculate area weights.}
}
\value{
a list of arguments to pass on to downstream functions from NIcalc.
}
\description{
This is a helper function which collates arguments used by other functions
from the NIcalc packages when calculating thematic and fully custom indices.
For thematic indices, only arguments "OutputType" and "theme" need to be
provided. For fully custom indices, other arguments need to be provided as
well.
}
\details{
The arguments specified here get passed on to several functions from the
NIcalc package, including assembleNiObject() and calculateNIunitWeight() and
we refer to the NIcalc documentation for more details.
}
