% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listIndicators_NIforEC.R
\name{listIndicators_NIforEC}
\alias{listIndicators_NIforEC}
\title{Make indicator list for a Nature Index to be used in ecological condition
assessments}
\usage{
listIndicators_NIforEC(ecosystem)
}
\arguments{
\item{ecosystem}{character string specifying the ecosystem of interest.
Currently supported are "Forest", "Mountain", and "Wetland".}
}
\description{
The Nature Index is used in the framework for ecological condition assessment
as an indicator of biodiversity (representing species abundance and trends).
To fulfill that role, the index for a specific ecosystem will not necessarily
be made up of the full set of indicators that are used in its standalone
implementation as part of the reporting of the Norwegian Nature Index.
This function writes ecosystem-specific lists of indicators for making
versions of the Nature Index for ecological condition assessments.
}
\details{
NOTES:

This function is incomplete and needs to be updated once discussions
about which indicators should be part of the different products have been
concluded (expected late 2023).

For now, I have implemented an alternative approach where indicator lists for
the Ecological Condition products are derived from the indicator set for the
corresponding Nature Index, minus a TBD set of indicators that get dropped.
}
