% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIflex-package.R
\docType{package}
\name{NIflex-package}
\alias{NIflex}
\alias{NIflex-package}
\title{NIflex: Flexible workflows for generating Nature Index (products)}
\description{
This package contains a series of wrapper-, helper-, and plotting functions that together make up a flexible workflow for downloading data from the database for the Norwegian Nature Index and using it to calculate the Nature Index for different ecosystems, pre-defined thematic indices to publish alongside the Nature Index, modified Nature Index products for use in Ecological Condition Assessments, and fully customizable Nature Index products.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NINAnor/NIviz}
}

}
\author{
\strong{Maintainer}: Chloé R. Nater \email{chloe.nater@nina.no} (\href{https://orcid.org/0000-0002-7975-0108}{ORCID})

}
\keyword{internal}
