% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NIunits_AcidEutr}
\alias{NIunits_AcidEutr}
\title{Spatial unit information for thematic indices Acidification and Eutrophication}
\format{
\subsection{\code{NIunits_AcidEutr}}{

A data frame with 240 rows and 3 columns:
\describe{
\item{BSunitID}{Integer ID for relevant BSunits}
\item{name}{Norwegian name of the BSunit}
\item{NIunitName}{Norwegian name of the associated NIunit}
...
}
}
}
\source{
R:/Prosjekter/12892000_naturindeks_rammeverk_database_og_innsynslosning/Naturindeks 2020/Beregninger/Tema ferskvann/themeEutrophication/NIområder eutrofiering forsuring.txt
}
\usage{
NIunits_AcidEutr
}
\description{
Data for re-defining spatial units (BSunits) for the thematic indicators
Acidification and Eutrophication.
}
\keyword{datasets}
