% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNI_DensityRidgeTS.R
\name{plotNI_DensityRidgeTS}
\alias{plotNI_DensityRidgeTS}
\title{Plot a density ridge time series for a custom index}
\usage{
plotNI_DensityRidgeTS(
  Index,
  OutputType,
  ecosystem = NULL,
  theme = NULL,
  allAreas = FALSE,
  selectedArea = "wholeArea"
)
}
\arguments{
\item{Index}{a list ontaining all information on the custom index. Object
"CustomIndex" in the output of calculateCustomNI().}

\item{OutputType}{character. The type of index to be plotted. Can be one of
c("NatureIndex", "EcologicalCondition", "ThematicIndex", "CustomIndex").}

\item{ecosystem}{character. The ecosystem of the index. Optional argument
required when OutputType = "NatureIndex" or "EcologicalCondition". Can be one
of c("Skog", "Fjell", "Våtmark", "Åpent lavland", "Ferskvann", "Kystvann",
"Hav"). Note that only the first three are relevant for OutputType =
"EcologicalCondition" so far.}

\item{theme}{character. Optional argument specifying which thematic index
should be plotted. Required when OutputType = "ThematicIndex". For currently
supported thematic indices, see documentation of listIndicators_thematicIndex().}

\item{allAreas}{logical. If FALSE (default), makes a single-panel plot for
the area specified via the argument "selectedArea". If TRUE, makes a multi-
panel plot with one panel per area available.}

\item{selectedArea}{character. Name of the area for which to plot. Can be any
of the list object names of argument "Index". The default is "wholeArea".}
}
\value{
a density ridge time-series plot.
}
\description{
This function was developed based on materials in the NIviz repository
(https://github.com/NINAnor/NIviz).
}
\details{
It is an alternative approach for visualizing time series of indicator
values including uncertainty. Using the "ggridges" package, the function
plots not just summary statistics but the entire index value distributions.
The color code is set to match the same scale as on the Nature Index webpage.
}
