% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNI_Map.R
\name{plotNI_Map}
\alias{plotNI_Map}
\title{Plot maps of custom index median, confidence interval, and displacement}
\usage{
plotNI_Map(
  shp,
  year,
  OutputType,
  ecosystem = NULL,
  theme = NULL,
  plotMedian = TRUE,
  plotCI = TRUE,
  plotDisplacement = FALSE,
  interactiveMap = FALSE
)
}
\arguments{
\item{shp}{an sf object containing geomapped summary statistics for the index in
the specified year.}

\item{year}{integer. Which year to plot maps for.}

\item{OutputType}{character. The type of index to be plotted. Can be one of
c("NatureIndex", "EcologicalCondition", "ThematicIndex", "CustomIndex").}

\item{ecosystem}{character. The ecosystem of the index. Optional argument
required when OutputType = "NatureIndex" or "EcologicalCondition". Can be one
of c("Skog", "Fjell", "Våtmark", "Åpent lavland", "Ferskvann", "Kystvann",
"Hav"). Note that only the first three are relevant for OutputType =
"EcologicalCondition" so far.}

\item{theme}{character. Optional argument specifying which thematic index
should be plotted. Required when OutputType = "ThematicIndex". For currently
supported thematic indices, see documentation of listIndicators_thematicIndex().}

\item{plotMedian}{logical. If TRUE (default), plots index median.}

\item{plotCI}{logical. If TRUE (default), plots index confidence interval.}

\item{plotDisplacement}{logical. If TRUE (default = FALSE) plots index
dislocation.}

\item{interactiveMap}{logical. If FALSE (default) plots a static map. If TRUE,
plots an interactive (zoomable) map.}
}
\value{
one or two static or interactive maps of index summary metrics.
}
\description{
This function allows to plot three types of summary metrics of calculated
custom indices onto a map: median, spread of the confidence interval (= a
measure of uncertainty), and the statistical displacement (an artifact
resulting from indicator scaling that may affect the index estimates.)
At present, the function can plot any of these separately, or in pairs of two
(but not all three simultaneously).
}
\details{
#' This function was developed based on materials in the NIviz repository
(https://github.com/NINAnor/NIviz).

Plotting to map requires shapefiles for the spatial units used by the Nature
Index for the different ecosystems and the different thematic indices.
These are deposited in NINA's internal file system under
"P:/41201611_naturindeks_2021_2023_vitenskapelig_ansvar/Shapefiles".
Alternatively, the files can be downloaded in JSON format from the NI
database API here: https://ninweb08.nina.no/NaturindeksAPI/index.html. Files
downloaded from here first need to be converted into shapefiles.
}
