% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNI_StandardTS.R
\name{plotNI_StandardTS}
\alias{plotNI_StandardTS}
\title{Plot area-specific and area-aggregated time series of calculated indices}
\usage{
plotNI_StandardTS(Index, addAverage = TRUE, truncateY = FALSE)
}
\arguments{
\item{Index}{a list ontaining all information on the custom index. Object
"CustomIndex" in the output of calculateCustomNI().}

\item{addAverage}{logical. If TRUE (default), a line for the area-averaged index is plotted together with area-specific estimates.}

\item{truncateY}{logical. If TRUE, the y axis of the plot will be truncated to only visualize the range of values relevant to the specific index. If FALSE (default), the y-axis will span the entire possible range from 0 to 1.}
}
\value{
a time series plot.
}
\description{
This function is an extended version of NIcalc::plot.NiSeries() which simultaneously plots indices calculated for multiple areas.
}
