% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDropIndices.R
\name{selectDropIndicators}
\alias{selectDropIndicators}
\title{Make a list of selected indicators that should be dropped from an index}
\usage{
selectDropIndicators(
  DropIndMode = "pre-defined",
  OutputType = NULL,
  ecosystem = NULL,
  customList = NULL
)
}
\arguments{
\item{DropIndMode}{character. If "pre-defined" (default), lists a default set
of indicators from indicator sets for Nature Index and Ecological Condition
outputs. If "custom", returns the same indicator list as provided through the
"customList" argument.}

\item{OutputType}{character. The type of output the workflow will create,
here an optional argument. Can be one of
c("NatureIndex", "EcologicalCondition", "ThematicIndex", "CustomIndex").}

\item{ecosystem}{character. The ecosystem for which to calculate the index.
Optional argument required when OutputType = "NatureIndex" or
"EcologicalCondition". Can be one of c("Skog", "Fjell", "Våtmark",
"Åpent lavland", "Ferskvann", "Kystvann", "Hav"). Note that only the first
three are relevant for OutputType = "EcologicalCondition" so far.}

\item{customList}{a vector containing integer IDs for indicators to drop.
Optional argument that needs to be provided when DropIndMode = "custom".}
}
\value{
a vector containing integer IDs for indicators to drop.
}
\description{
This function gives some more control over indices with pre-defined sets of
indicators.
}
\details{
When DropIndMode is set to "pre-defined", the list of indicators to drop will
be a default set that is specified within the function. At present, this
includes the indicators "Lavhei" in forest and wetlands ecosystems and "Sei"
in ocean ecosystem. These have previously been dropped from the calculation
of the Nature Index for the relevant ecosystem was they were deemed little
representative for the ecosystem.
This list may be extended somewhat when updating the Nature Index in 2025 as
there are some additional indicators that may have to be dropped due to
insufficient data for updating.

When DropIndMode is set to "custom", one may pass a list of any length
containing the integer IDs of indicators one wishes to drop from calculations.
The indicator IDs can currently only be found in the Nature Index database
(https://naturindeks.nina.no/) and login credentials are required. It would
be good to display indicator IDs publicly, or make it possible to supply
indicator names to this function instead in the future.
}
